requirements/base.txt
Django
Pillow
django-extensions
django-filter
django-taggit
djangorestframework
django-cors-headers